+++ CHANGES.rst
@@ -6,7 +6,7 @@ Version 2.1.0
 Unreleased
 
 -   Update Click dependency to >= 8.0.
-
+-   Added support for Decimal objects in JSONEncoder
 
 Version 2.0.2
 -------------